requirements.txt
gunicorn==19.4.5
Django==1.8.8
Unipath==1.0
dj-static==0.0.6
python-decouple==2.2
